Rename tokenize.py to tokenization.py
Avoids a collision with standard lib tokenize module; not sure if this
is causing our weird coverage issues
