+++ gufe/network.py
@@ -4,7 +4,7 @@
 from typing import FrozenSet, Iterable, Optional, Tuple
 
 import networkx as nx
-from .tokenize import GufeTokenizable
+from .tokenization import GufeTokenizable
 
 from .chemicalsystem import ChemicalSystem
 from .transformations import Transformation
