+++ gufe/transformations/transformation.py
@@ -4,7 +4,7 @@
 from typing import Optional, Iterable
 
 from openff.toolkit.utils.serialization import Serializable
-from ..tokenize import GufeTokenizable
+from ..tokenization import GufeTokenizable
 
 from ..chemicalsystem import ChemicalSystem
 from ..protocols import Protocol, ProtocolDAG, ProtocolResult, ProtocolDAGResult
